requirements.txt
# Core environment
python-dotenv>=1.0.1
scikit-learn>=1.5.0

# LangChain ecosystem
langchain>=0.2.11
langchain-core>=0.2.24
langchain-community>=0.2.8
langchain-openai>=0.1.15
langchain-pinecone>=0.1.1
langchain-text-splitters>=0.2.1

# Pinecone client
pinecone-client>=3.2.0

# DocArray (for in-memory search vector store)
docarray>=0.40.0
